requirements.txt
# Core dependencies for the Cloze Fine-tuning Project
# Python >= 3.8 required

# Deep Learning and NLP
torch>=1.8.1
transformers>=4.4.2
datasets>=1.5.0

# Data processing
pandas>=1.2.0
numpy>=1.20.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Utilities
PyYAML>=5.4.0
tqdm>=4.60.0

# Optional: For advanced text processing
# nltk>=3.6.0
# spacy>=3.2.0

# Optional: For logging and monitoring
# wandb>=0.12.0
# tensorboard>=2.6.0

# Development dependencies (optional)
# pytest>=6.2.0
# black>=21.0.0
# flake8>=3.9.0